Scaffolding for other game states
I tend to like developing in a way that "fleshes out" the
prototype/early stages so that way I feel more motivated as it seems
more obvious that the game is in more "ready" stages.

I won't have art for a while so this is extra important for my
motivation right now.