Stage_State system (read more)
The Stage_State will be intended to store the choreography for a level, and should be either based off a lua script or
will be hardcoded through a state machine in game code.

It's technically intended to be a little separate from the game code as it's more of a scripted sequence.