Scaffolded enemy / bullet behaviors for tomorrow's actual gameplay.
Bullets have a concept of what they originate from to determine what they should be damaging. Enemies are basically just smarter bullets,
but otherwise will work identically.

Since this is a more casual bullet hell (because I'm not a danmaku pro), my encounter design isn't going to be super hard, but I'll try to not make it too easy.