Stages have been moved to coroutines successfully
So this is a nasty bit of business since it uses a crap ton of macros, but it
does provide a nicer high level interface for defining levels in C++ when I need to.

The goal is to just use lua, but it's not a big deal to write some in C++ and then translate it
later.