fixed software renderer tiled clipping problem
NOTE: I am aware of an acne problem for rotated sprites. I know what is causing it, however it's not a big enough deal for me to care about
right now. It is due to the fact that the renderer will write into a different tile which causes the acne.