[DEMOSYS,need:clean up,integration,ui] Handled some general behavior cases with the system interacting with the rest of the game.
The system is basically "finalized" since it works by doing frame recordings of input. So at least it's not going to change. Like ever.

I just mainly need a UI and handling specific rules around "continues", of which I can just insert a "continue" frame... if it's needed.

Although obviously, the intention is to record levels. Or full stages even... Basically the core is complete. It's just some semantics that need,
some adjustment.