[GAME] Added UI atlas, trying to add trophy icons and utilize the nine patch UI to make things look nicer
This is one of the days, where I'm going to try to improve the "polish"/look of the game, although there's not
many serious code changes. Technically this is a content workload.

I'll do more tonight, but I'm not expecting to accomplish as much this week as I thought I would :/ Which is fine, just keep going.