[EMSCRIPTEN BUILD]: Modified engine code to compile emscripten build, verify that the engine indeed compiles.
I cannot verify that the engine works, but if it's anything like legends, I hope it doesn't explode terribly once I try to
get some stuff kind of working... It's a bit annoying but hopefully it should work fine...

NOTE: building for emscripten is theoretically second class and not something I'm genuinely interested in anymore beyond having
it as a gimmick.

Target platform is PC windows almost certainly even if I use SDL2.