Added renderer change disclaimer to restart
The renderer swapping is... not stable as soon as you try to switch
between hardware renderers (D3D11 -> OpenGL, OpenGL -> D3D11), which
is not really a common thing to do anyway (probably the device
contexts fighting over the same GPU or something?), but I'd prefer to
get rid of anything that's easily crashing.

While I could make it stable, the code for this is uh a mess and it's
not super important, but I'd rather cover up anything that goes wrong anyway.