[RENDER-WOBBLY-TEXT] wobbly text has been added as an engine-native ability
I don't actually believe this is a good architectural system, in
reality I should've had a push glyph primitive exposed, but
render_commands do not expose this, and this is a *faster* way of
doing it.

It utilizes the flag field and does the wobbly logic by hooking into
the GameUI's wobbly factor which is a stupid amount of coupling, but
I'm not very concerned with code quality right now...

Fortunately it's a really easy change to track down so it's not a big
deal to be honest.

The logo also wobbles according to the GameUI wobble settings so it
looks relatively consistent and pretty nice.

I certainly could've made a much better decision, but this is fine for
now since it's not a particularly invasive change in my
opinion. Although definitely in the future, I should've just stuck to
what I did in LEGENDS where it was just a game-specific feature that
wasn't hooked into the engine like it is right now.