beginning to add burst meter addition
This is a new gameplay mechanic idea I've drafted and it involves
utilizing a meter that will build-up and be consumed upon focus.

It's like an attack stamina bar basically.

The specification is mostly this commit message.

It is designed to change the behaviors of offensive actions during
"focus-mode" to give the player more tools to either build score or
survive a wave better.

It is meant to build up:
   - Passively when NOT in focus mode, this is the slowest build-up
   - When an enemy is killed (most)
   - When an enemy is hit (some)
   - Faster when grazing a bullet
It is used when:
   - Remaining in focus mode (very light)
   - Attacking in focus mode (extra, changes based on tier, should increase as you attack longer)
     it is intended to force the player to only use "burst" shots as
     otherwise they will lose burst and be more disadvantaged.
   - Bomb! (drain all burst, and perform different action based on tier)
   - Any "death" will empty the entire meter.
It should:
   Change the way attacking works for 3 different tiers, which emulate
   a "power level".

   Allow for 3 different bomb abilities. The final level is the
   classic bomb (score gaining wipeout abilities)

NOTE: to balance things out, the bomb action cannot occur without at least 1 tier of burst mode.

It can be used when:
   - BURST has not just been drained.
   - If it has been drained, it cannot be used until it has charged up
     to the "second tier" of usage.
   - If a bomb drains the burst meter, you must wait until the "first tier" of usage
     NOTE: this will also remove any benefits of FOCUS mode except for
     the more precise movement. (so no attack modifications, and no bomb usage)

I did want to allow a passive "bullet-time" on high focus modes but
that might not be super trivial to-do without breaking massive amounts
of stuff, but it *is* in consideration...

I'll probably need to tune this *a-lot* as stages are about 2-3
minutes so there need to be at least 2 chances per stage to reliably
use the bomb.