[DEMO SETUP] Demo build targets + preprocessor flag
I want to have a demo build ready by the end of the month (and announce
a few days after NEXT fest).

My timing is not that great, but there's not really a lot of good time
in Q2 25 anyway.

If I can get the demo parts done well enough in advance, I can do chill
for a little bit.