[USERPREFPATH] Shuffled around some code to get proper user preferences paths
I am also now using SDL2 for that (which is why get_preference_directory
was moved from common.cpp to main.cpp, as common.cpp is also linked by
Bigfilepacker which does not use SDL2.)

The win32 code was correct, but I'm using SDL2 as the platform layer so
I might as well be consistent for that as I do not know what the APIs
for all the various operating systems might be.

Yeah... I really gotta fix that in the future.